Add files via upload
Using a regression model, this project aims to predict the fair rental price of apartments in São Paulo. The steps include data collection, preprocessing, training the regression model, evaluation, and price prediction. The dataset is analyzed, cleaned, and split into training and testing sets. The model is fitted to the training data and evaluated using metrics such as MAE and MSE. Based on the trained model, accurate rental price predictions can be made. This project provides a practical solution for estimating the fair rental price in a specific real estate region.